Changed fetch of api-posts from app to postslist
