+++ code/public/index.html
@@ -3,7 +3,7 @@
 
   <head>
     <meta charset="utf-8" >
-    <meta name="viewport" content="width=device-width, initial-scale=1" >
+    <meta name="viewport" content="width=device-width, initial-scale=1 maximum-scale-1" >
     <meta property="og:title" content="Happy Thoughts">
     <meta property="og:description" content="A happy thoughts feed">
     <meta property="og:image" content="https://happy-thoughts-feed.netlify.app/assets/happy-feed.png">
